Refactor persistence layer: use IF NOT EXISTS for holdings table & move holdings logic into PortfolioDAO
- Updated UserDAO#createTables() to remove DROP TABLE and use CREATE TABLE IF NOT EXISTS user_holdings () so existing data isn’t lost on startup
- Refactored PortfolioDAO to inject IDatabaseConnection, remove in-memory constructor and implement `getHoldings()`, upsertHolding(), getAvailableBalance() and `getTotalPortfolioValue() directly against the `user_holdings / user_balances tables
- Adjusted PortfolioController to use the new PortfolioDAO methods for loading and displaying user holdings

+++ src/main/java/com/javarepowizards/portfoliomanager/models/PortfolioEntry.java
@@ -13,7 +13,7 @@ public class PortfolioEntry {
 
     public StockName getStock() { return stock; }
     public double getPurchasePrice() { return purchasePrice; }
-    public double getAmountHeld() { return amountHeld; }
+    public int getAmountHeld() { return (int) amountHeld; }
 
     public double getBuyPrice() {return purchasePrice;}
     public double getQuantity() {return amountHeld;}
